move to resources_utils
