+++ sky/clouds/aws.py
@@ -441,7 +441,7 @@ class AWS(clouds.Cloud):
 
     def _get_feasible_launchable_resources(
         self, resources: 'resources_lib.Resources'
-    ) -> 'resources_lib.FeasibleResources':
+    ) -> resources_utils.FeasibleResources:
         if resources.instance_type is not None:
             assert resources.is_launchable(), resources
             # Check the instance type is valid in the cloud
@@ -452,10 +452,10 @@ class AWS(clouds.Cloud):
                 region=resources.region,
                 zone=resources.zone)
             if not regions:
-                return resources_lib.FeasibleResources([], [], None)
+                return resources_utils.FeasibleResources([], [], None)
             # Treat Resources(AWS, p3.2x, V100) as Resources(AWS, p3.2x).
             resources = resources.copy(accelerators=None)
-            return resources_lib.FeasibleResources([resources], [], None)
+            return resources_utils.FeasibleResources([resources], [], None)
 
         def _make(instance_list):
             resource_list = []
@@ -481,9 +481,9 @@ class AWS(clouds.Cloud):
                 memory=resources.memory,
                 disk_tier=resources.disk_tier)
             if default_instance_type is None:
-                return resources_lib.FeasibleResources([], [], None)
+                return resources_utils.FeasibleResources([], [], None)
             else:
-                return resources_lib.FeasibleResources(
+                return resources_utils.FeasibleResources(
                     _make([default_instance_type]), [], None)
 
         assert len(accelerators) == 1, resources
@@ -499,10 +499,10 @@ class AWS(clouds.Cloud):
             zone=resources.zone,
             clouds='aws')
         if instance_list is None:
-            return resources_lib.FeasibleResources([], fuzzy_candidate_list,
+            return resources_utils.FeasibleResources([], fuzzy_candidate_list,
                                                    None)
-        return resources_lib.FeasibleResources(_make(instance_list),
-                                               fuzzy_candidate_list, None)
+        return resources_utils.FeasibleResources(_make(instance_list),
+                                                           fuzzy_candidate_list, None)
 
     @classmethod
     @functools.lru_cache(maxsize=1)  # Cache since getting identity is slow.
