+++ sky/clouds/ibm.py
@@ -266,12 +266,12 @@ class IBM(clouds.Cloud):
 
     def _get_feasible_launchable_resources(
         self, resources: 'resources_lib.Resources'
-    ) -> 'resources_lib.FeasibleResources':
+    ) -> 'resources_utils.FeasibleResources':
         fuzzy_candidate_list: List[str] = []
         if resources.instance_type is not None:
             assert resources.is_launchable(), resources
             resources = resources.copy(accelerators=None)
-            return resources_lib.FeasibleResources([resources],
+            return resources_utils.FeasibleResources([resources],
                                                    fuzzy_candidate_list, None)
 
         def _make(instance_list):
@@ -297,9 +297,9 @@ class IBM(clouds.Cloud):
                 memory=resources.memory,
                 disk_tier=resources.disk_tier)
             if default_instance_type is None:
-                return resources_lib.FeasibleResources([], [], None)
+                return resources_utils.FeasibleResources([], [], None)
             else:
-                return resources_lib.FeasibleResources(
+                return resources_utils.FeasibleResources(
                     _make([default_instance_type]), [], None)
 
         assert len(accelerators) == 1, resources
@@ -314,9 +314,9 @@ class IBM(clouds.Cloud):
             zone=resources.zone,
             clouds='ibm')
         if instance_list is None:
-            return resources_lib.FeasibleResources([], fuzzy_candidate_list,
+            return resources_utils.FeasibleResources([], fuzzy_candidate_list,
                                                    None)
-        return resources_lib.FeasibleResources(_make(instance_list),
+        return resources_utils.FeasibleResources(_make(instance_list),
                                                fuzzy_candidate_list, None)
 
     @classmethod
