+++ sky/clouds/lambda_cloud.py
@@ -178,12 +178,12 @@ class Lambda(clouds.Cloud):
 
     def _get_feasible_launchable_resources(
         self, resources: 'resources_lib.Resources'
-    ) -> 'resources_lib.FeasibleResources':
+    ) -> 'resources_utils.FeasibleResources':
         if resources.instance_type is not None:
             assert resources.is_launchable(), resources
             # Accelerators are part of the instance type in Lambda Cloud
             resources = resources.copy(accelerators=None)
-            return resources_lib.FeasibleResources([resources], [], None)
+            return resources_utils.FeasibleResources([resources], [], None)
 
         def _make(instance_list):
             resource_list = []
@@ -209,9 +209,9 @@ class Lambda(clouds.Cloud):
                 memory=resources.memory,
                 disk_tier=resources.disk_tier)
             if default_instance_type is None:
-                return resources_lib.FeasibleResources([], [], None)
+                return resources_utils.FeasibleResources([], [], None)
             else:
-                return resources_lib.FeasibleResources(
+                return resources_utils.FeasibleResources(
                     _make([default_instance_type]), [], None)
 
         assert len(accelerators) == 1, resources
@@ -227,9 +227,9 @@ class Lambda(clouds.Cloud):
             zone=resources.zone,
             clouds='lambda')
         if instance_list is None:
-            return resources_lib.FeasibleResources([], fuzzy_candidate_list,
+            return resources_utils.FeasibleResources([], fuzzy_candidate_list,
                                                    None)
-        return resources_lib.FeasibleResources(_make(instance_list),
+        return resources_utils.FeasibleResources(_make(instance_list),
                                                fuzzy_candidate_list, None)
 
     @classmethod
