+++ sky/clouds/scp.py
@@ -251,16 +251,16 @@ class SCP(clouds.Cloud):
 
     def _get_feasible_launchable_resources(
         self, resources: 'resources_lib.Resources'
-    ) -> 'resources_lib.FeasibleResources':
+    ) -> 'resources_utils.FeasibleResources':
         # Check if the host VM satisfies the min/max disk size limits.
         is_allowed = self._is_disk_size_allowed(resources)
         if not is_allowed:
-            return resources_lib.FeasibleResources([], [], None)
+            return resources_utils.FeasibleResources([], [], None)
         if resources.instance_type is not None:
             assert resources.is_launchable(), resources
             # Accelerators are part of the instance type in SCP Cloud
             resources = resources.copy(accelerators=None)
-            return resources_lib.FeasibleResources([resources], [], None)
+            return resources_utils.FeasibleResources([resources], [], None)
 
         def _make(instance_list):
             resource_list = []
@@ -287,9 +287,9 @@ class SCP(clouds.Cloud):
                 memory=resources.memory,
                 disk_tier=resources.disk_tier)
             if default_instance_type is None:
-                return resources_lib.FeasibleResources([], [], None)
+                return resources_utils.FeasibleResources([], [], None)
             else:
-                return resources_lib.FeasibleResources(
+                return resources_utils.FeasibleResources(
                     _make([default_instance_type]), [], None)
 
         assert len(accelerators) == 1, resources
@@ -305,9 +305,9 @@ class SCP(clouds.Cloud):
             zone=resources.zone,
             clouds='scp')
         if instance_list is None:
-            return resources_lib.FeasibleResources([], fuzzy_candidate_list,
+            return resources_utils.FeasibleResources([], fuzzy_candidate_list,
                                                    None)
-        return resources_lib.FeasibleResources(_make(instance_list),
+        return resources_utils.FeasibleResources(_make(instance_list),
                                                fuzzy_candidate_list, None)
 
     @classmethod
