+++ sky/utils/resources_utils.py
@@ -9,6 +9,7 @@ from typing import List, Optional, Set
 from sky.utils import ux_utils
 
 if typing.TYPE_CHECKING:
+    from sky import resources as resources_lib
     from sky import backends
 
 _PORT_RANGE_HINT_MSG = ('Invalid port range {}. Please use the format '
@@ -157,3 +158,21 @@ def get_readable_resources_repr(handle: 'backends.CloudVmRayResourceHandle',
                                            launched_resource_str)
             return f'{handle.launched_nodes}x {launched_resource_str}'
     return _DEFAULT_MESSAGE_HANDLE_INITIALIZING
+
+
+@dataclasses.dataclass
+class FeasibleResources:
+    """Feasible resources returned by cloud.
+
+    Used to represent a collection of feasible resources returned by cloud,
+    any fuzzy candidates, and optionally a string hint if no feasible resources
+    are found.
+
+    Fuzzy candidates example: when the requested GPU is A100:1 but is not
+    available in a cloud/region, the fuzzy candidates are results of a fuzzy
+    search in the catalog that are offered in the location. E.g.,
+    ['A100-80GB:1', 'A100-80GB:2', 'A100-80GB:4', 'A100:8']
+    """
+    resources_list: List['resources_lib.Resources']
+    fuzzy_candidate_list: List[str]
+    hint: Optional[str]
